FollowButtonの責任範囲をfollow/unfollowのみに
