+++ src/pages/Entity.tsx
@@ -1,6 +1,6 @@
-import { Box, Button, Paper, Tab, Tabs, Typography, alpha, useTheme } from '@mui/material'
+import { Box, Button, IconButton, Paper, Tab, Tabs, Typography, Zoom, alpha, useTheme } from '@mui/material'
 import { useEffect, useMemo, useRef, useState } from 'react'
-import { useParams } from 'react-router-dom'
+import { Navigate, useParams } from 'react-router-dom'
 import { useApi } from '../context/api'
 import type { Character, Entity, StreamElementDated } from '../model'
 import type { Userstreams } from '../schemas/userstreams'
@@ -10,6 +10,8 @@ import { CCAvatar } from '../components/CCAvatar'
 import { Timeline } from '../components/Timeline'
 import { useObjectList } from '../hooks/useObjectList'
 import Background from '../resources/defaultbg.png'
+import InfoIcon from '@mui/icons-material/Info'
+import CreateIcon from '@mui/icons-material/Create'
 import { FollowButton } from '../components/FollowButton'
 
 export function EntityPage(): JSX.Element {
@@ -19,10 +21,17 @@ export function EntityPage(): JSX.Element {
     const [entity, setEntity] = useState<Entity>()
     const [profile, setProfile] = useState<Character<Profile>>()
     const [streams, setStreams] = useState<Character<Userstreams>>()
+    const [mode, setMode] = useState<'info' | 'edit'>('info')
     const messages = useObjectList<StreamElementDated>()
     const scrollParentRef = useRef<HTMLDivElement>(null)
+    const isSelf = id === api.userAddress
 
     const [tab, setTab] = useState(0)
+
+    const transitionDuration = {
+        enter: theme.transitions.duration.enteringScreen,
+        exit: theme.transitions.duration.leavingScreen
+    }
 
     useEffect(() => {
         if (!id) return
@@ -88,7 +97,56 @@ export function EntityPage(): JSX.Element {
                 >
                     <b>{profile.payload.body.username || 'anonymous'}</b>
                 </Button>
-                <FollowButton userCCID={id} />
+                <Box
+                    sx={{
+                        position: 'relative',
+                        width: '40px',
+                        height: '40px',
+                        mr: '8px'
+                    }}
+                >
+                    {isSelf ? (
+                        <>
+                            <Zoom
+                                in={mode === 'info'}
+                                timeout={transitionDuration}
+                                style={{
+                                    transitionDelay: `${mode === 'info' ? transitionDuration.exit : 0}ms`
+                                }}
+                                unmountOnExit
+                            >
+                                <IconButton
+                                    sx={{ p: '8px', position: 'absolute' }}
+                                    onClick={() => {
+                                        setMode('edit')
+                                    }}
+                                >
+                                    <CreateIcon sx={{ color: 'primary.contrastText' }} />
+                                </IconButton>
+                            </Zoom>
+                            <Zoom
+                                in={mode === 'edit'}
+                                timeout={transitionDuration}
+                                style={{
+                                    transitionDelay: `${mode === 'edit' ? transitionDuration.exit : 0}ms`
+                                }}
+                                unmountOnExit
+                            >
+                                <IconButton
+                                    sx={{ p: '8px', position: 'absolute' }}
+                                    onClick={() => {
+                                        setMode('info')
+                                    }}
+                                >
+                                    <InfoIcon sx={{ color: 'primary.contrastText' }} />
+                                </IconButton>
+                            </Zoom>
+                            {mode === 'edit' && <Navigate to="/settings" />}
+                        </>
+                    ) : (
+                        <>{id && <FollowButton userCCID={id} />}</>
+                    )}
+                </Box>
             </Box>
             <Box /* body */
                 sx={{
